Changed new.lwjgl.org to www.lwjgl.org
The domain new.lwjgl.org for the license link does not appear to work. This may just be me, but I would not think so. I asked in the IRC and got no response a while back, so I just decided to send this. Being that this is a new website and that may have been done in a transitioning period, I am assuming the subdomain was removed but this file was not changed. Plus, http://www.lwjgl.org/licence does in fact contain a BSD-style license, as was described, so I believe this is okay.

Changed http://new.lwjgl.org/license to http://www.lwjgl.org/license on line 223.