Route middleware for nprogress + System.import
By wrapping the route we allow smarter use of nprogress ( e.g. not
showing it on first load since the view is rendered on the server
already ).
We also allow webpack to tree shake more efficiently
Finally, we don't require an additional file per route
index.js is now the route itself