Fix container height in px to avoid *layout*
Android re-calculates vh height for video container as soon as the
address bar is hidden. This resulted in page reflow and pushed all
the content a few pixels down. By setting a fixed px height we avoid
it.

TODO: When Aphrodite stops outputting !important we can then go back
to using as single className and avoid the componentDidMount() hack