fix: nest descendant subscriptions, fire when !sCU
We nest subscriptions of descendant components, similar to what
react-redux is doing in ./src/utils/Subscription.js, so that we can
ensure the ancestor components re-render before descendants.

This avoid bugs like removing a TODO item from the store while the
component rendering that TODO is still mounted. Previous implementation
fired TODO's listener before the ancestor had the chance to unmount it.