feat: use pagehide instead of unload event to store scroll location
It's recommended to always use the pagehide event to detect possible page
unloads (a.k.a the terminated state) rather than the unload event. Many
developers treat the unload event as a guaranteed callback and use it as an
end-of-session signal to save state and send analytics data, but doing this is
extremely unreliable, especially on mobile! The unload event does not fire in
many typical unload situations, including closing a tab from the tab switcher
on mobile or closing the browser app from the app switcher.

More information here:
https://developers.google.com/web/updates/2018/07/page-lifecycle-api